l2hub-vote-bot/requirements.txt

discord-py
python-dotenv
flask
aiohttp

